Fixes #336 where StartingEvent never got dispatched.
This bug was caused by the dispatcher arg for shardimpl never being
set from StatelessBot in _create_shard. I have made that argument
mandatory now to prevent this in the future.

I have noticed that ShardClientImpl has now got a _components
as well as a _dispatcher field. This shouldn't have been added as
it is effectively redundant information. Thus, I have created
#337 to address this.