Removed entity and re-enabled slotting on objects.
This had been disabled by accident to stop errors caused by multiple
inheritance with slots triggering class layout errors on initialization
of the classes.

The solution has been to remove the Entity entirely. I took this
chance to remove Unique and put it with the Snowflake class instead.
Each class that has an app or an ID now explicitly declares this.

This reduces the memory usage for some models by more than 50% in some
cases, so this is a pretty big deal. It also reduces the time complexity
for any attribute access.