Relicenced future versions to use more permissive MIT license.
This will enable us in future to bundle other libraries that may be used to
improve performance, such as similar-licensed C extensions. In addition, it
will allow the project to grow and become available to a wider audience.

As I am the sole copyright owner of the source by each license header, the
source remains under my control to relicense.

Please contact me if you have any queries with regards to this.