Adjusted error handling of events a little.
If at least one listener to an ExceptionEvent exists, then it will now be logged
as a DEBUG level log entry.

If no listeners to ExceptionEvent instances exist, then it will be logged as an
ERROR level log entry as expected.

Also removed insomnia temporarily, as I plan to change it a bit.