Optimised raw event dispatching to uncover bug.
Looks like, at least on my machine, asyncio immediately invokes anything you await
rather than switching to another task on the queue first unless the call does raw
IO. I have confirmed this with Epoll, Poll and Select selector implementations on
a non-debug asyncio SelectorEventLoop implementation.

This means that the bulk of dispatching an event would currently occur as soon as
the event is dispatched rather than after another task runs, which could lead to
immediate slowdown if other tasks are queued.

Switching to sync dispatching and using create task to invoke the callback
management "later" seems to speed up this implementation significantly and allows
other race conditions we have not accounted for properly as part of #77 to be
detectable with test scripts that saturate the event loop.