Implemented session start limit detection.
Attempting to start more sessions than you are allowed will now result
in a GatewayError on startup before any shards are initialized.

Being wary that Discord may or may not have bodged this somehow, or not
bothered to override the responses from this endpoint for larger bots
that need higher limits, I have added an extra argument to bot.run
and bot.start: `ignore_session_start_limit`, which will enable the user
to skip this check entirely.