Fixed another shard bug where zombieing blocks on receive websocket calls.
We now wait for several seconds, give up listening to the websocket, recheck
the heartbeat state, and then continue again repeatedly. This means that if
we are unable to receive anything, we give up shortly to detect zombification.

This is done as it is a little tidier than creating more tasks to do the
same thing.