Added new logic to shard that should hopefully detect random socket closures.
It appears Windows will send an EOF to the TCP socket if an adapter closes.
This seems to happen when connecting to a VPN or putting the computer into
hibernation/sleep mode. When this occurs, aiohttp assumes it means a normal
closure and gracefully closes, but we do not usually have a way to determine
this.

We can however see that aiohttp invokes ClientWebSocketResponse#close which
is the same call we make. The solution is to make a facading method that we
invoke that sets a flag determining that we did in fact want to close. If this
doesn't get set, we know aiohttp is the one to blame.