Improved handling of bucket ratelimits in edge cases.
Running two bots on the same token should now not error
if you get ratelimited due to tracking issues.

The example case we found was running two instances of
the same bot that responded to pings with a specific
reaction under the same token. This should now simply
log a message to say that this has happened.

Co-authored-by: nekokatt <nekoka.tt@outlook.com>