Implemented default max-retry of 60s.
This means if any REST rate limit occurs that lasts more than 60 seconds,
you will instead get a hikari.errors.RateLimitTooLongError get raised
instead which contains details of how long to back off for should you
wish to retry later.

You can disable this functionality by passing `max_rate_limit=float("inf")`
to the `hikari.impl.bot.BotApp` and `hikari.impl.rest.RESTApp` constructors
if you have a reason to not want this. The assumption is that in 99% of
cases, waiting for sustained periods of time would be confusing to the
user and lead to a poorer user experience overall, which is why this has
been implemented.

I still need to fix the tests.